chore: extract image substitutors to the image package
